Davide/docs setup (#517)
A few changes to the build workflow done on my fork that look to be working so can be added to the ACCESS-NRI repo'. Note that there was an extreme amount of hacking done to `.github/workflows/deploy_to_github_pages.yml`. (@atteggiani will fix this tomorrow/get PR previews working etc.)